+++ src/main/java/roomescape/domain/reservation/repository/impl/ReservationTimeDAO.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.repository.impl;
+package roomescape.domain.reservation.repository.impl;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -16,8 +16,8 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.common.exception.AlreadyInUseException;
 import roomescape.common.exception.EntityNotFoundException;
-import roomescape.reservation.entity.ReservationTime;
-import roomescape.reservation.repository.EntityRepository;
+import roomescape.domain.reservation.entity.ReservationTime;
+import roomescape.domain.reservation.repository.EntityRepository;
 
 @Repository
 public class ReservationTimeDAO implements EntityRepository<ReservationTime> {
