+++ src/main/java/roomescape/ui/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.admin.controller;
+package roomescape.ui;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,5 +14,15 @@ public class AdminController {
     @GetMapping("/admin")
     public String dashboard() {
         return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String adminReservationDashboard() {
+        return "/admin/reservation";
+    }
+
+    @GetMapping("/admin/time")
+    public String adminReservationTimeDashboard() {
+        return "admin/time";
     }
 }
