+++ src/test/java/roomescape/reservation/entity/ReservationTimeTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import roomescape.domain.reservation.entity.ReservationTime;
 
 class ReservationTimeTest {
